requirements.txt
# CTI Project - Requirements
# Core dependencies
requests>=2.31.0
psycopg2-binary>=2.9.9
redis>=5.0.1
python-dotenv>=1.0.0
schedule>=1.2.0
PyYAML>=6.0.1

# Data processing
pandas>=2.1.4
numpy>=1.26.0
python-dateutil>=2.8.2

# Web scraping
beautifulsoup4>=4.12.2
lxml>=4.9.3
feedparser>=6.0.10

# OpenCTI integration
pycti>=6.0.0
stix2>=3.0.1
stix2-patterns>=2.0.0

# Threat Intelligence APIs
shodan>=1.31.0
OTXv2>=1.5.12
vt-py>=0.18.2

# Communication
python-telegram-bot>=20.7

# Monitoring
prometheus-client>=0.19.0
structlog>=23.2.0

# Excel generation
openpyxl==3.1.2
xlsxwriter==3.1.2

# Database
sqlalchemy==2.0.19

# Utilities
click>=8.1.7
rich>=13.7.0
tweepy>=4.0.0
tabulate>=0.9.0

# Async utilities
aiofiles>=23.1.0
aiohttp>=3.9.0

# Note: asyncio-timeout is not needed for Python 3.11+
# It's part of the standard library as asyncio.timeout